pyproject.toml
[project]
name = "dspy-multimodal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dspy>=3.0.3",
    "pillow>=12.0.0",
]
